fix: move logging setup to hyperopter
restores hyperopt logging functionality
